requirements.txt
aiodns==3.0.0
aiohttp==3.8.5
aiosignal==1.2.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.11.0
beautifulsoup4==4.12.2
bleach==4.1.0
blis==0.7.9
Bottleneck==1.3.5
Brotli==1.0.9
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.0.4
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.2
confection==0.1.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
executing==0.8.3
fastai==2.7.13
fastapi==0.103.0
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.16.2
fastkaggle==0.0.8
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.12.2
gmpy2==2.1.2
h11==0.12.0
huggingface_hub==0.20.2
idna==3.4
ipykernel==6.25.0
ipython==8.20.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
kaggle==1.6.3
kiwisolver==1.4.4
langcodes==3.3.0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.0
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.4
munkres==1.1.4
murmurhash==1.0.7
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.1
notebook==7.0.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.26.3
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
parso==0.8.3
Pillow==10.0.1
pip==23.3.1
platformdirs==3.10.0
ply==3.11
preshed==3.0.6
prometheus-client==0.14.1
prompt-toolkit==3.0.43
psutil==5.9.0
pure-eval==0.2.2
pycares==4.0.0
pycparser==2.21
pydantic==1.10.12
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.30.2
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rpds-py==0.10.6
safetensors==0.4.1
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
setuptools==68.2.2
shellingham==1.5.0
sip==6.7.12
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.8
stack-data==0.2.0
starlette==0.27.0
sympy==1.12
terminado==0.17.1
text-unidecode==1.3
thinc==8.2.2
threadpoolctl==2.2.0
timm==0.9.12
tinycss2==1.2.1
toml==0.10.2
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.3.3
tqdm==4.65.0
traitlets==5.7.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.20.0
wasabi==0.9.1
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
websocket-client==0.58.0
wheel==0.41.2
win-inet-pton==1.1.0
yarl==1.9.3
